requirements.txt
fastapi
uvicorn
pydantic
pdfplumber
python-docx
pdf2image
pytesseract
tiktoken
langchain
langchain-community
pinecone
python-multipart
openai